Added Token service, cleaned up session service

+++ src/main/java/dev/startupstack/tenantservice/services/external/FirebaseSDKService.java
@@ -2,6 +2,7 @@ package dev.startupstack.tenantservice.services.external;
 
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.util.Optional;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
@@ -19,33 +20,56 @@ import com.google.firebase.auth.FirebaseToken;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.jboss.logging.Logger;
 
-import dev.startupstack.tenantservice.Constants;
-
 /**
  * FirebaseSDKService
  */
-//@Singleton
 @Dependent
 public class FirebaseSDKService {
 
     private static final Logger LOG = Logger.getLogger(FirebaseSDKService.class);
 
-
     @Inject
     @ConfigProperty(name = "startupstack.tenantservice.firebase.keyfile")
-    //TODO: Figure out why this injection is not working
-    static String serviceAccountFile = "/home/jargelo/firebase-sa.key.json";
+    String serviceAccountFile;
 
-    public static FirebaseToken verifyToken(String token) throws WebApplicationException {
+    public Optional<String> verifyToken(String accessToken) throws WebApplicationException {
+        Optional<String> result = Optional.empty();
         try {
-            return FirebaseAuth.getInstance().verifyIdToken(token);
-        } catch (FirebaseAuthException exception) {
-            throw new WebApplicationException(exception.getMessage());
+            FirebaseAuth.getInstance().verifyIdToken(accessToken);
+            return result;
+        } catch (FirebaseAuthException fae) {
+            result = Optional.of("Token validation failed: " + fae.getMessage());
+            LOG.warn(result.get());
+            return result;
+        } catch (IllegalArgumentException iae) {
+            throw new WebApplicationException(iae.getMessage(), iae);
         }
     }
 
-    public static void initialize() {
-        LOG.info(serviceAccountFile);
+    public FirebaseToken getDecryptedToken(String accessToken) {
+        try {
+            return FirebaseAuth.getInstance().verifyIdToken(accessToken, true);
+        } catch (FirebaseAuthException fae) {
+            LOG.warn("Unable to get token: " + fae.getMessage());
+            return null;
+        } catch (IllegalArgumentException iae) {
+            throw new WebApplicationException(iae.getMessage(), iae);
+        }
+    }
+
+    
+    public void revokeTokens(String uid) throws WebApplicationException {
+        try {
+            LOG.infof("[%s] Revoking tokens ...", uid);
+            FirebaseAuth.getInstance().revokeRefreshTokens(uid);
+            LOG.infof("[%s] Revoke tokens OK", uid);
+        } catch (IllegalArgumentException | FirebaseAuthException exception) {
+            throw new WebApplicationException(exception.getMessage(), exception);
+        }
+    }
+
+    @PostConstruct
+    void initialize() {
         if (FirebaseApp.getApps().isEmpty()) {
             try {
                 FileInputStream serviceAccount = new FileInputStream(serviceAccountFile);
@@ -61,7 +85,9 @@ public class FirebaseSDKService {
 
     @PreDestroy
     void predestroy() {
-        FirebaseApp.getInstance().delete();
+        if (FirebaseApp.getApps().isEmpty() == false) {
+            FirebaseApp.getInstance().delete();    
+        }
     }
     
 }
